SP-StockBot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
line-bot-sdk==3.5.0
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
openpyxl==3.1.5
pandas==2.2.3
pyarrow==17.0.0
groq==0.9.0
tenacity==8.4.2
apscheduler==3.10.4
matplotlib==3.9.2
orjson==3.10.7
python-dotenv==1.0.1
psutil==6.1.0
